move `ExecuteTimings` from `runtime::bank` to `program_runtime::timings`
