+++ core/Cargo.toml
@@ -43,6 +43,7 @@ solana-metrics = { path = "../metrics", version = "=1.10.0" }
 solana-net-utils = { path = "../net-utils", version = "=1.10.0" }
 solana-perf = { path = "../perf", version = "=1.10.0" }
 solana-poh = { path = "../poh", version = "=1.10.0" }
+solana-program-runtime = { path = "../program-runtime", version = "=1.10.0" }
 solana-rpc = { path = "../rpc", version = "=1.10.0" }
 solana-replica-lib = { path = "../replica-lib", version = "=1.10.0" }
 solana-runtime = { path = "../runtime", version = "=1.10.0" }
